Moved Metada class to common file. It works in Python2 and Python3
