--- linkero/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-
-class Metadata:
-    def __init__(self):
-        self.__version__ = '0.9.1'
-        self.__author__ = 'Rubén de Celis Hernández'
-
-    def get_version(self):
-        return self.__version__
-    def get_author(self):
-        return self.__author__
